Extract `gmath_nodes` from `gcore::ops`, create mods `debug` and `extract_xy`
