+++ editor/src/messages/portfolio/document/node_graph/node_properties.rs
@@ -12,7 +12,7 @@ use graph_craft::Type;
 use graph_craft::document::value::TaggedValue;
 use graph_craft::document::{DocumentNode, DocumentNodeImplementation, NodeId, NodeInput};
 use graphene_std::animation::RealTimeMode;
-use graphene_std::ops::XY;
+use graphene_std::extract_xy::XY;
 use graphene_std::path_bool::BooleanOperation;
 use graphene_std::raster::curve::Curve;
 use graphene_std::raster::{
@@ -1840,7 +1840,7 @@ pub fn offset_path_properties(node_id: NodeId, context: &mut NodePropertiesConte
 }
 
 pub fn math_properties(node_id: NodeId, context: &mut NodePropertiesContext) -> Vec<LayoutGroup> {
-	use graphene_std::ops::math::*;
+	use graphene_std::math_nodes::math::*;
 
 	let document_node = match get_document_node(node_id, context) {
 		Ok(document_node) => document_node,
