+++ editor/src/messages/portfolio/document_migration.rs
@@ -24,6 +24,51 @@ const REPLACEMENTS: &[(&str, &str)] = &[
 	("graphene_core::ConstructArtboardNode", "graphene_core::graphic_element::ToArtboardNode"),
 	("graphene_core::ToGraphicElementNode", "graphene_core::graphic_element::ToElementNode"),
 	("graphene_core::ToGraphicGroupNode", "graphene_core::graphic_element::ToGroupNode"),
+	("graphene_core::ops::MathNode", "graphene_math_nodes::MathNode"),
+	("graphene_core::ops::AddNode", "graphene_math_nodes::AddNode"),
+	("graphene_core::ops::SubtractNode", "graphene_math_nodes::SubtractNode"),
+	("graphene_core::ops::MultiplyNode", "graphene_math_nodes::MultiplyNode"),
+	("graphene_core::ops::DivideNode", "graphene_math_nodes::DivideNode"),
+	("graphene_core::ops::ModuloNode", "graphene_math_nodes::ModuloNode"),
+	("graphene_core::ops::ExponentNode", "graphene_math_nodes::ExponentNode"),
+	("graphene_core::ops::RootNode", "graphene_math_nodes::RootNode"),
+	("graphene_core::ops::LogarithmNode", "graphene_math_nodes::LogarithmNode"),
+	("graphene_core::ops::SineNode", "graphene_math_nodes::SineNode"),
+	("graphene_core::ops::CosineNode", "graphene_math_nodes::CosineNode"),
+	("graphene_core::ops::TangentNode", "graphene_math_nodes::TangentNode"),
+	("graphene_core::ops::SineInverseNode", "graphene_math_nodes::SineInverseNode"),
+	("graphene_core::ops::CosineInverseNode", "graphene_math_nodes::CosineInverseNode"),
+	("graphene_core::ops::TangentInverseNode", "graphene_math_nodes::TangentInverseNode"),
+	("graphene_core::ops::RandomNode", "graphene_math_nodes::RandomNode"),
+	("graphene_core::ops::ToU32Node", "graphene_math_nodes::ToU32Node"),
+	("graphene_core::ops::ToU64Node", "graphene_math_nodes::ToU64Node"),
+	("graphene_core::ops::ToF64Node", "graphene_math_nodes::ToF64Node"),
+	("graphene_core::ops::RoundNode", "graphene_math_nodes::RoundNode"),
+	("graphene_core::ops::FloorNode", "graphene_math_nodes::FloorNode"),
+	("graphene_core::ops::CeilingNode", "graphene_math_nodes::CeilingNode"),
+	("graphene_core::ops::MinNode", "graphene_math_nodes::MinNode"),
+	("graphene_core::ops::MaxNode", "graphene_math_nodes::MaxNode"),
+	("graphene_core::ops::ClampNode", "graphene_math_nodes::ClampNode"),
+	("graphene_core::ops::EqualsNode", "graphene_math_nodes::EqualsNode"),
+	("graphene_core::ops::NotEqualsNode", "graphene_math_nodes::NotEqualsNode"),
+	("graphene_core::ops::LessThanNode", "graphene_math_nodes::LessThanNode"),
+	("graphene_core::ops::GreaterThanNode", "graphene_math_nodes::GreaterThanNode"),
+	("graphene_core::ops::LogicalOrNode", "graphene_math_nodes::LogicalOrNode"),
+	("graphene_core::ops::LogicalAndNode", "graphene_math_nodes::LogicalAndNode"),
+	("graphene_core::ops::LogicalNotNode", "graphene_math_nodes::LogicalNotNode"),
+	("graphene_core::ops::BoolValueNode", "graphene_math_nodes::BoolValueNode"),
+	("graphene_core::ops::NumberValueNode", "graphene_math_nodes::NumberValueNode"),
+	("graphene_core::ops::PercentageValueNode", "graphene_math_nodes::PercentageValueNode"),
+	("graphene_core::ops::CoordinateValueNode", "graphene_math_nodes::CoordinateValueNode"),
+	("graphene_core::ops::ColorValueNode", "graphene_math_nodes::ColorValueNode"),
+	("graphene_core::ops::GradientValueNode", "graphene_math_nodes::GradientValueNode"),
+	("graphene_core::ops::StringValueNode", "graphene_math_nodes::StringValueNode"),
+	("graphene_core::ops::DotProductNode", "graphene_math_nodes::DotProductNode"),
+	("graphene_core::ops::SizeOfNode", "graphene_core::debug::SizeOfNode"),
+	("graphene_core::ops::SomeNode", "graphene_core::debug::SomeNode"),
+	("graphene_core::ops::UnwrapNode", "graphene_core::debug::UnwrapNode"),
+	("graphene_core::ops::CloneNode", "graphene_core::debug::CloneNode"),
+	("graphene_core::ops::ExtractXyNode", "graphene_core::extract_xy::ExtractXyNode"),
 	("graphene_core::logic::LogicAndNode", "graphene_core::ops::LogicAndNode"),
 	("graphene_core::logic::LogicNotNode", "graphene_core::ops::LogicNotNode"),
 	("graphene_core::logic::LogicOrNode", "graphene_core::ops::LogicOrNode"),
